Refactored sorce code structure.
