+++ NeoMathEngine/src/CPU/x86/avx/CMakeLists.txt
@@ -9,14 +9,18 @@ target_sources(${PROJECT_NAME}
 
     # Sources
     ./src/AvxMathEngine.cpp
-    ./src/MatrixMultiplying/AvxMatrixMultiplying.cpp
+    ./src/MatrixMultiplyingInterleaved/AvxMatrixMultiplying.cpp
     # Headers
     ./common.h
     ./src/BlobConvolution.h
     ./src/BlobConvolutionImpl.h
     ./src/AvxCommon.h
-    ./src/MatrixMultiplying/Interleavers/Interleavers.h
-    ./src/MatrixMultiplying/Kernels/Kernels_6x.h
+    ./src/MatrixMultiplyingInterleaved/Interleavers/Interleavers.h
+    ./src/MatrixMultiplyingInterleaved/MicroKernels/Kernel_AVX_6x16.h
+    ./src/MatrixMultiplyingInterleaved/MicroKernels/Kernel_AVX_6x8.h
+    ./src/MatrixMultiplyingInterleaved/MicroKernels/Kernel_AVX_6x4.h
+    ./src/MatrixMultiplyingInterleaved/MicroKernels/Kernel_AVX_6x2.h
+    ./src/MatrixMultiplyingInterleaved/MicroKernels/Kernel_AVX_6x1.h
 )
 
 string(TOUPPER ${CMAKE_SYSTEM_NAME} UPPERCASE_CMAKE_SYSTEM_NAME)
@@ -28,9 +32,9 @@ target_include_directories(${PROJECT_NAME}
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../..>
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../../MatrixMultiplyingInterleavedCommon>
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/src>
-        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/src/MatrixMultiplying>
-        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/src/MatrixMultiplying/Interleavers>
-        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/src/MatrixMultiplying/Kernels>
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/src/MatrixMultiplyingInterleaved>
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/src/MatrixMultiplyingInterleaved/Interleavers>
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/src/MatrixMultiplyingInterleaved/MicroKernels>
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../../../../include>
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../../..>
 )
